Playerlist will now update after delete / enable confirmations
Moved logic into player list pinia store, so multiple items can update the list
