+++ client/src/components/players/PlayerList.vue
@@ -1,46 +1,16 @@
 <script setup lang="ts">
 
 import {onMounted, ref, watch} from 'vue';
-import axios from "axios";
 import PlayerTile from "@/components/players/Tiles/PlayerTile.vue";
-import type {PlayerListItem} from "@/components/players/Objects/Player";
 import InputText from "primevue/inputtext";
-import type {Log} from "@/components/players/Objects/ActivityLogs";
-import type {ChangedProperty} from "@/components/players/Objects/ChangedProperty";
+import {playerList} from "@/components/players/Stores/PlayerListStore";
+const playerListStore = playerList();
 
-let players = ref<Array<PlayerListItem>>([]);
-const filteredPlayers = ref<Array<PlayerListItem>>([]);
 const searchQuery = ref<string>("");
 
-function fetchData() {
-  axios.get('/admin/users')
-      .then((response) => {
-        response.data.users.forEach(function(item:PlayerListItem) {
-          item.lockedOutExpires = new Date(item.lockedOutExpires);
-        });
-        players.value = response.data.users;
-        filteredPlayers.value = response.data.users
-      });
-}
-
-onMounted(() =>{
-  fetchData();
+onMounted(async () =>{
+  await playerListStore.fetchPlayers();
 })
-
-function filterPlayers(query: string) {
-  const lowercasedQuery = query.toLowerCase().trim();
-
-  if (!lowercasedQuery) {
-    // Reset showing all players if the query is empty
-    filteredPlayers.value = players.value;
-  } else {
-    // Filter players by username or email
-    filteredPlayers.value = players.value.filter((player) =>
-        player.username.toLowerCase().includes(lowercasedQuery) ||
-        player.email.toLowerCase().includes(lowercasedQuery)
-    );
-  }
-}
 
 // Debounce function
 function debounce(fn: Function, delay: number) {
@@ -55,7 +25,7 @@ function debounce(fn: Function, delay: number) {
 
 // Debounced filter function
 const debouncedFilterPlayers = debounce((query: string) => {
-  filterPlayers(query);
+  playerListStore.filterPlayers(query);
 }, 250);
 
 // Watch for changes to the search query and trigger the debounced filter function
@@ -81,11 +51,11 @@ watch(searchQuery, (newQuery) => {
         />
       </div>
     </div>
-    <div v-if="filteredPlayers.length === 0" class="m-3">
+    <div v-if="playerListStore.filteredPlayers.length === 0" class="m-3">
       No users with that name or email address
     </div>
     
-    <div v-for="player in filteredPlayers" :key="player.id">
+    <div v-for="player in playerListStore.filteredPlayers" :key="player.id">
       <PlayerTile :player-info="player" />
     </div>
   </div>
